Change int test to float

+++ workout.py
@@ -38,8 +38,8 @@ def exercise():
 
     # Server-side checks to ensure Set One values aren't null
     try:
-        set_one_reps_trial = int(request.form.get(f"set-one-reps"))
-        set_one_weight_trial = int(request.form.get(f"set-one-weight"))
+        set_one_reps_trial = float(request.form.get(f"set-one-reps"))
+        set_one_weight_trial = float(request.form.get(f"set-one-weight"))
     except ValueError:
         return "ERROR"
